cropProject/requirements.txt
Django==2.1.1
django-bom==1.30
djangorestframework==3.9.2
scikit-image==0.14.0
scikit-learn==0.20.0
pandas==0.23.4
numpy==1.15.2
matplotlib==2.2.2
mysqlclient==1.3.12
requests==2.21.0